+++ jamanchu/src/main/java/com/recipe/jamanchu/auth/jwt/JwtFilter.java
@@ -1,7 +1,7 @@
 package com.recipe.jamanchu.auth.jwt;
 
 import com.recipe.jamanchu.entity.UserEntity;
-import com.recipe.jamanchu.model.dto.request.UserDetailsDTO;
+import com.recipe.jamanchu.model.dto.request.auth.UserDetailsDTO;
 import com.recipe.jamanchu.model.type.UserRole;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
