+++ jamanchu/src/main/java/com/recipe/jamanchu/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.recipe.jamanchu.controller;
 
-import com.recipe.jamanchu.model.dto.request.SignupDTO;
+import com.recipe.jamanchu.model.dto.request.auth.SignupDTO;
 import com.recipe.jamanchu.model.type.ResultCode;
 import com.recipe.jamanchu.service.UserService;
 import lombok.RequiredArgsConstructor;
