+++ jamanchu/src/main/java/com/recipe/jamanchu/service/UserService.java
@@ -1,6 +1,6 @@
 package com.recipe.jamanchu.service;
 
-import com.recipe.jamanchu.model.dto.request.SignupDTO;
+import com.recipe.jamanchu.model.dto.request.auth.SignupDTO;
 import com.recipe.jamanchu.model.type.ResultCode;
 
 public interface UserService {
